+++ packages/ui/src/primitives/forms/DayTimePicker.js
@@ -63,7 +63,6 @@ export const DateTimePicker = (props: Props) => {
     yearPickerType,
   } = props;
   const TODAY = new Date();
-  console.log(offset);
   return (
     <div id={htmlID}>
       <DayPicker