+++ packages/core/CHANGELOG.md
@@ -1,6 +1,7 @@
 # @voussoir/core
 
 ## 1.1.0
+
 - [patch] 8d8666ad:
 
   - Dependency upgrade: graphql -> 14.0.3, graphql-tools -> 4.0.3